functionning merge
